Addressed review comments, separated ScreenCapturerManager in it's own class file and exposed more declarative APIs. Check for build version before ScreenCapturerManager is used to avoid foreground start/stop service when it is not necessary
